requirements.txt
asgiref==3.8.1
certifi==2025.1.31
chardet==3.0.4
Django==5.0.4
django-ckeditor==6.7.2
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.14.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pytz==2025.1
redis==5.0.1
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.1
